Merging r142659: ------------------------------------------------------------------------ r142659 | baldrick | 2011-10-21 10:04:00 -0700 (Fri, 21 Oct 2011) | 6 lines
Errors here should be ignored, as they are already for all other
rules in this Makefile.  Otherwise the nightly test fails if clang
(or dragonegg) fails to compile some test, because this rule then
fails to convert the .ll output from clang (which doesn't exist
because clang failed) to .bc.

------------------------------------------------------------------------


git-svn-id: https://llvm.org/svn/llvm-project/test-suite/branches/release_30@142662 91177308-0d34-0410-b5e6-96231b3b80d8