[test-suite] Prevent timeit/fpcmp from being rebuilt every time make is run.
This patch adds a custom_command level underneath the custom_targets for timeit
and fpcmp to prevent the executables from being rebuilt every time make is run.

The custom_target rule will still be run every time, but the binaries will be
detected as being up to date and skip rebuilding.

http://reviews.llvm.org/D15938

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@257059 91177308-0d34-0410-b5e6-96231b3b80d8