[test-suite] Add ClangAnalyzer directory to test-suite repository
Add a new directory in the llvm test-suite repository to hold test projects for
the static analyzer. This directory will be used by a Green Dragon build bot
(discussed in http://lists.llvm.org/pipermail/cfe-dev/2015-September/045199.html
that will periodically re-analyze the projects specified in this directory to
detect when analysis results change unexpectedly.

Rather than check in the source code for these projects, we instead check in (1)
scripts to download, cache, and build the project and (2) expected reference
results in plist form from running the analyzer over the project. There are
pre-existing scripts in utils/analyzer in the clang project that understand the
layout of this directory (see http://reviews.llvm.org/D14345).

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@258336 91177308-0d34-0410-b5e6-96231b3b80d8