Add cmake caches for common configurations
There are some common configurations of the test-suite which really
should be encoded right into the test-suite so they can be easily used
and shared.

This commit adds new cmake caches to start to capture those configs.
Release is a production build where we enable -O3 and LTO.  Debug, is
what a developer might use on their desktop while working on code, with
the flags -O0 -g.

Watch for a matching LNT commit which adds support for using these.

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@263182 91177308-0d34-0410-b5e6-96231b3b80d8