lit: Big refactoring, introduce TestPlan object
The testplan object describes the sequence of steps that will get
executed to run the benchmark and collect the metrics. This allows for a
nicer separation of different aspects into modules.

For example the codesize and compiletime module can add a metric collection
step to the testplan, the profilegen module can modify all commandline
of the runscript and add a later step for merging the profile data. All
these aspects can be added with a unified 'module.mutatePlan()' from test.py.

This also unifies a bunch of code paths that deal with error handling
and reporting so we should properly display errors in metric collection
steps now.

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@263532 91177308-0d34-0410-b5e6-96231b3b80d8