[LNT] Update timeit.sh to have the same return code as timeit when --append-exitstatus is used
Summary:
When passing --append-exitstatus, timeit will write the exit status
to the output and return 0 (with some exceptions in some edge cases).
The exit status will then be checked as part of the output with fpcmp.

We need timeit.sh to do the same (otherwise we will wrongly record a
failure for some tests).

Reviewers: jmolloy

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D20361

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@269912 91177308-0d34-0410-b5e6-96231b3b80d8