[test-suite] Use shellcommand.wrap() when adding the run_under command
Summary:
When mutating a command like "cd X; ./cmd" to add the run_under
arguments (for example "taskset -c 1"), the mutated command needs to
maintain the "cd X;" part in the output command prefix:
   "cd X; taskset -c 1 ./cmd"

If we don't do this, the mutated command cannot be parsed.

Use shellcommand.wrap() to mutate the command instead of string
concatenation, since shellcommand knows how to handle the
"cd X; ..".

Reviewers: MatzeB

Subscribers: llvm-commits

Differential Revision: http://reviews.llvm.org/D20854

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@271661 91177308-0d34-0410-b5e6-96231b3b80d8