nbench: Eliminate timing and re-running related code
nbench is by far the noisiest benchmark in the test-suite for me. This
seems to be caused by the benchmark running a variable number of
iterations internally based on its own measurements. Even though the
original commit to the test-suite claims this was disabled.

This commit shortcuts all the re-running until confident code and remove
all time measurement code just to be sure.

The number of iterations is also reduced to 1 to reduce the overal
test-suite runtime (see the discussion in the review).

Differential Revision: http://reviews.llvm.org/D21590

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@273529 91177308-0d34-0410-b5e6-96231b3b80d8