[test-suite] Supress warnings of missing reference output files.
Summary:
For the Bitcode tests, there is a large number of tests that do not have or need a reference file, leading to a warning for each test.
Yet, it's still cleaner to use the same methods in SingleMultiSource.cmake which trigger the warning.
Add a new variable NO_REFERENCE_OUTPUT, checked in SingleMultiSource.cmake
before setting the reference file. Set this variable to supress the warnings.
Alternative suggestions on addressing this are welcome.

Reviewers: llvm-commits, MatzeB, mehdi_amini

Differential Revision: https://reviews.llvm.org/D22548

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@276131 91177308-0d34-0410-b5e6-96231b3b80d8