[test-suite] Bitcode tests: Update cmake to build driver and halide runtime only once.
Summary:
The simd_ops bitcode tests used to build the common driver and the halide runtime for each test.
Update cmake so they are build just once.
Note that in the future the halide runtime could be taken up one directory level and used in common for
multiple tests and benchmark. However building just the one .bc file into a static library failed on OSX.

Reviewers: mehdi_amini, MatzeB

Subscribers: llvm-commits

Differential Revision: https://reviews.llvm.org/D23127

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@277684 91177308-0d34-0410-b5e6-96231b3b80d8