[polybench] check against strict fp
This patch makes each test in polybench to run twice: once with the flags
specified by the user, and once with strict FP computations: optnone and
fp-contract=off.  The results are then compared in memory with FP_ABSTOLERANCE.
If there are differences, the program stops returning 1.  The output of the
strict FP computation is printed in a file and compared against a hashed
reference output.

With this patch the polybench will pass when the test-suite is compiled with
-ffp-contract=on.

Differential Revision: https://reviews.llvm.org/D25346

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@284334 91177308-0d34-0410-b5e6-96231b3b80d8