Bitcode: Enable single_source lit config
The compile_time collection fails for the bitcode tests because the test
executable name has no relation to the bitcode input.

Setting single_source=True is not enough to fix it, but at least it
stops lit from summing up the compile of every tests every time (which
slowed the testrunner noticably down).

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@295274 91177308-0d34-0410-b5e6-96231b3b80d8