Set higher timeouts for PAQ8p and SciMark2-C on AArch64
When running the test-suite on AArch64 at -O0, we sometimes get timeouts
in these 2 benchmarks. To avoid this, we increase the thresholds in the
Makefiles. For SciMark2-C, which takes between 450-500s, we use a
threshold of 750s, and for PAQ8p, which takes between 650-750s, we use
1000.

We're modifying the Makefiles because this is what the AArch64 buildbots
currently use.

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@305671 91177308-0d34-0410-b5e6-96231b3b80d8