[lit, test-suite] Fix a reference to FileBasedTest that I missed earlier.
Summary:
The test-suite implements its own specialized lit format for running tests,
which was based on FileBasedTest. Since that base was removed, the format now
needs to use something that is still available.

The only functionality that comes from FileBasedTest is the test discovery,
which was moved into ShTest. For now, it seems to make the most sense simply to
use ShTest as the base for TestSuiteTest.

I think a better solution is to factor out the test discovery logic, although
that will require a bit more work (and has effects across projects).

Reviewers: leandron

Subscribers: llvm-commits, sanjoy

Differential Revision: https://reviews.llvm.org/D34830

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@306730 91177308-0d34-0410-b5e6-96231b3b80d8