cmake/xcode_sdk: Work around problems with ranlib/strip
ranlib/strip in LTO builds require libLTO.dylib to read the bitcode
embedded in the .o/.a files. However by default they pick up the library
coming with the SDK while we usually want to pick up the one coming with
the compiler (in case we are not using the compiler coming with the
SDK). Create shims for ranlib/strip to avoid this problem.

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@309077 91177308-0d34-0410-b5e6-96231b3b80d8