[test-suite] Adding the miniAMR Benchmark
miniAMR applies a stencil calculation on a unit cube computational domain,
which is divided into blocks. The blocks all have the same number of cells in
each direction and communicate ghost values with neighboring blocks. With
adaptive mesh refinement, the blocks can represent different levels of
refinement in the larger mesh. Neighboring blocks can be at the same level or
one level different, which means that the length of cells in neighboring blocks
can differ by only a factor of two in each direction. The calculations on the
variables in each cell is an averaging of the values in the chosen stencil. The
refinement and coarsening of the blocks is driven by objects that are pushed
through the mesh. If a block intersects with the surface or the volume of an
object, then that block can be refined. There is also an option to uniformly
refine the mesh. Each cell contains a number of variables, each of which is
evaluated indecently.

Web: https://mantevo.org/packages/
Github: https://github.com/Mantevo/miniAMR

When run on Intel Xeon E5-2699 v4 @ 2.20GHz:
compile_time: 33.7526
exec_time: 0.4879
Maximum resident set size (kbytes): 836784

Patch by Brian Homerding, thanks!

Includes glibc_compat_rand.{h,c} written by me.

Differential Revision: https://reviews.llvm.org/D36682

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@311398 91177308-0d34-0410-b5e6-96231b3b80d8