test-suite: add avx512 tests with math intrinsics
Summary:
Here is the bunch of tests. In each test we do some operations with intrinsics and also check result.
This set depends on cpu features, so it needs support to identify avx512 features in cmake and makefile approaches of test-suite. Here are patches for makefile and cmake: [[https://reviews.llvm.org/D38182|D38182]], [[https://reviews.llvm.org/D38484|D38484]]


Reviewers: craig.topper, zvi, MatzeB

Reviewed By: MatzeB

Subscribers: llvm-commits, mgorny

Differential Revision: https://reviews.llvm.org/D41249

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@321144 91177308-0d34-0410-b5e6-96231b3b80d8