[ubsan] Make the C++ code here not cast floating point values outside the range of an integer type to that integer type.
Instead, change the random value function to cast to the destination
type first and do the devision by 8 afterward. When the destination is
floating point this should be roughly equivalent, but when it is an
integer type it will end up truncating stuff down rather than ending up
outside the range.

Differential Revision: https://reviews.llvm.org/D46295

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@331240 91177308-0d34-0410-b5e6-96231b3b80d8