[test-suite] Fix UB in initialization code for stepanov_container.
This benchmark wants to use vector::operator[] to form a pointer-past-the-end.
This is not valid to do via operator[], because doing so requires forming an
(intermediate) reference-past-the-end (dangling references are always UB, simply
by existing). Forming address-past-the-end is valid under pointer arithmetic
(basic.compound p3.2) and iterators (iterator.requirements.general p7), however.

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@331314 91177308-0d34-0410-b5e6-96231b3b80d8