[test-suite][ubsan] Remove optimization that causes UB in SimpleMOC.
This patch replaces the hand written ceil() with a call to ceilf().  This small
optimization is outside the primary kernel and should show little to no
difference in performance for the test suite.  Previously the UB was only used
in an '==' comparison for an early exit from the loop and had no effect on the
program behavior.


git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@332395 91177308-0d34-0410-b5e6-96231b3b80d8