[test-suite] Allow --use-perf=profile and --run-under to work together
When the command line mutation for adding perf commands runs it first
tries to find the results of the run-under command line mutation. In a
recent refactoring the run-under mutateCommandLine function was renamed
to _mutateCommandLine which resulted in a 'module' object has no attribute
error.

This change renames _mutateCommandLine back to mutateCommandLine as it
isn't private to the run-under command.



git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@335463 91177308-0d34-0410-b5e6-96231b3b80d8