[XRay][test-suite] Benchmarks for profiling mode implementation
Summary:
This patch adds microbenchmarks for the XRay Profiling Mode
implementation to the test-suite.

The benchmarks included cover:

- Cost of the Profiling Mode runtime handler(s) and underlying
  implementation details when enabled.

- Different benchmarks for different call stack traces. Initially
  showing deep, shallow, and wide function call stacks.

These microbenchmarks can be used to measure progress on the
optimisation work associated with the profiling mode runtime
implementation going forward. It also allows us to better qualify the
cost of the XRay runtime framework (in particular the trampolines) as we
make improvements to those in the future.

Depends on D48653.

Reviewers: kpw, eizan

Reviewed By: eizan

Subscribers: mgorny, llvm-commits

Differential Revision: https://reviews.llvm.org/D48879

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@336970 91177308-0d34-0410-b5e6-96231b3b80d8