[CUDA, test-suite] Added test cases for the integet SIMD math functions.
CUDA-9.2 made these function compiler builtins, so clang no longer has access to
their implementation and we had to provide our own. These tests verify that the
result of these function match the ones provided by the reference
implementation.

Effectively the tests with CUDA-9.1 or older tests the reference
implementation. THe SIMD functions are provided by CUDA SDK and are the ground
truth for the test. The tests make sure that the code provided by the tests
matches it. The tests with CUDA-9.2 use clang-provided SIMD functions and the
implementation provided by the test is the ground truth.

Differential Revision: https://reviews.llvm.org/D49582

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@337570 91177308-0d34-0410-b5e6-96231b3b80d8