[test-suite, CUDA] Update CUDA cmake files.
This should unbreak CUDA buildbots.
The problem is that CUDA test suite is generating multiple
test executables that all use the same reference output file
and that resulted in multiple compilation jobs attempting to
create the same symlink at the same time.

This patch adds a SUFFIX parameter to make it possible to append
a suffix to the target file name and that allows us to avoid name clashes.

Differential Revision: https://reviews.llvm.org/D51663

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@341430 91177308-0d34-0410-b5e6-96231b3b80d8