[test-suite][mips] Disable setlocale calls in consumer-typeset test
When running the test-suite through qemu an error can occur when the
host and target architecture differ in endianness. This happens when the
test calls setlocale. The glibc implementation of this call reads
/usr/lib/locale/locale-archive as raw bites. Since the endianness
doesn't match, the numbers read from the archive don't make sense and
cause the check to fail.

This patch disables the setlocale call so the test passes regardless of
endianness.

Patch by Miloš Stojanović

Differential Revision: http://reviews.llvm.org/D55708

git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@350066 91177308-0d34-0410-b5e6-96231b3b80d8