Add geomean summary row to output from utils/compare.py.
This change adds a summary row with the geometric mean of the selected
metric to the output of utils/compare.py. The 'rhs' values are
normalized to the 'lhs' values and the geometric mean of the results
is computed.

This should be similar to the geomean row in the LNT HTML UI and allow
to compare 2 sets of runs.

Reviewers: anemet, MatzeB, cmatthews, serge-sans-paille

Reviewed By: anemet

Differential Revision: https://reviews.llvm.org/D57828


git-svn-id: https://llvm.org/svn/llvm-project/test-suite/trunk@356545 91177308-0d34-0410-b5e6-96231b3b80d8